Separate legacy resources out into `legacy_resources` so the code is easier to manage later
