+++ python_modules/libraries/dagster-airbyte/dagster_airbyte/__init__.py
@@ -27,14 +27,16 @@ from dagster_airbyte.asset_defs import (
 )
 from dagster_airbyte.ops import airbyte_sync_op as airbyte_sync_op
 from dagster_airbyte.resources import (
-    AirbyteCloudResource as AirbyteCloudResource,
     AirbyteCloudWorkspace as AirbyteCloudWorkspace,
-    AirbyteResource as AirbyteResource,
     AirbyteWorkspace as AirbyteWorkspace,
-    airbyte_cloud_resource as airbyte_cloud_resource,
-    airbyte_resource as airbyte_resource,
     load_airbyte_asset_specs as load_airbyte_asset_specs,
     load_airbyte_cloud_asset_specs as load_airbyte_cloud_asset_specs,
+)
+from dagster_airbyte.legacy_resources import (
+    AirbyteResource as AirbyteResource,
+    AirbyteCloudResource as AirbyteCloudResource,
+    airbyte_cloud_resource as airbyte_cloud_resource,
+    airbyte_resource as airbyte_resource,
 )
 from dagster_airbyte.translator import (
     AirbyteConnectionTableProps as AirbyteConnectionTableProps,
