+++ python_modules/libraries/dagster-airbyte/dagster_airbyte_tests/test_ops.py
@@ -4,8 +4,7 @@ from base64 import b64encode
 import pytest
 import responses
 from dagster import job, op
-from dagster_airbyte import AirbyteOutput, airbyte_resource, airbyte_sync_op
-from dagster_airbyte.resources import AirbyteCloudResource
+from dagster_airbyte import AirbyteOutput, airbyte_resource, airbyte_sync_op, AirbyteCloudResource
 
 DEFAULT_CONNECTION_ID = "02087b3c-2037-4db9-ae7b-4a8e45dc20b1"
